requirements.txt
annotated-types==0.5.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
ecdsa==0.18.0
enum-compat==0.0.3
future==0.18.3
idna==3.4
intuit-oauth==1.2.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pyasn1==0.5.0
pydantic==2.2.0
pydantic_core==2.6.0
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stripe==5.5.0
typing_extensions==4.7.1
urllib3==2.0.4
